Fix zhs for the putty/minitty settings
The problem was not that after fix ac34a0d the history was still wrong.
That fix only changed to not use not printable characters.
This one found one variable that used not printable characters and
clobber the zsh history when using ctrl-R.

This finally fixes that.
Still there is stuff to do:
    - detect UTF-8 and use special characters
    - make the prompt function more compact

One useful observation is to use print -P $PROMPT to debug prompt.