Fix the prompt ctrl-R
The prompt was broken by unseen characters that didn't worked on some
terminals(putty, mintty).
This was done by me, when I tried to make the prompt work without the
poor man lines drawing.
Unfortunately the PR_NO_COLOR is necessary to get from terminfo.
Now it is back to way it should work.

Funny though the tmux or screen fixes terminfo even if on borken
terminal.

The terminals are still somewhat mysterious to me, but less and less.