Hack space as leader and make it visible in showcmd
This is 'hack' to have space as leader. But make it visible in showcmd
window. This is due the fact that default ada plugins editor are soo
cool that they manually map some their stuff using leader. Which is cool
but they do it using 'exec' and string concatenation.
The problem is that leader is interpreted as space which is not added to
the mapping. This caused havoc for me when I tried to enter al or aj
(like in Mutu*al*_exclusion). So I found this wonderful hack that does
the trick.

Set leader to some innocent key like _
Then map your new super leader (<Space>) to the old leader.
So
map <Space> <Leader>
To make it work do not do:
    - nnoremap
    - put extra character (e.g space) after last character in mapping

This will do the trick and when you type '<Space>' then '_' will be shown in
showcmd window. The small section in bottom to show current 'command
buffer'.