Remove modifiable from quick fix settings
This can cause some grief, but quick fix is not good to modify.
Deleting the line will screw up the navigation and can make you jump to
incorrect error.

I know that you don't have to edit the line, just make them blank, but
it is easier to make it not modifiable.