Small vim key bindings for mutt
This is some small attempt to vimize mutt.
I have taken a way to use j,k as next,previous line in messages.
This is should be really convenient and to move to next message I should
use only one keystroke more (shift+ j or k).
Time will tell if that is the way to go.

On the upside gg and G is working as bottom/top message.

The idea is not entirely mine and it is stolen from
https://github.com/holomorph/dotfiles/blob/master/.config/mutt/bindings.muttrc