Fix SetMakeprg logic
The setlocal statements lacked the return after setting the makeprg.
Because of that the makeprg was not set correctly if there were makefile
present and you opened the .c file. As the if for '*.c' later in the
if's then it was chosen.

I also added rule to not override manual setting of makeprg.
If there is not default makeprg ('make') then leave it alone.