Fix keyboard settings on Debian 8 and xfce4
I want to have Polish letters and caps lock work as ctrl.
This is how i managed to do it.
The sys-configs/etc-default-keyboard is actually
/etc/default/keyboard file.
The .xsession-not is actually .xsession but I don't know the command to
start xfce4 session, and having it prohibits xfce4 to start.
The content of that file is important it have simmple command to add
custom keybinding from .xkb/symbols/local to existing keybindings.

The .Xmodmap have to go because it doesn't work with xkb.