Shorten up the if statement
The if in shell can be used without [] or test command.
So here i use grep and it's option -q to check for true and false of
this execution.

This is litte bit sad but i have just realized that
true evaluates to errno 0 and false to errno 1.
Should have known this by now, having written that many lines of 'C'.